First iteration on a basic HTTP relay server. (#1)
* First iteration on a basic HTTP relay server.

Mostly functional. Lot of flags and parameters to add and wire up.
The main machinery is made of three concurrent cooperating components (namely, an HTTP server, a master, N workers and a durable store).

The HTTP server persist incoming request (for durability) after some basic validation / normalization.

A master component is in charge of scanning and scheduling requests whereas workers are merely responsible for request delivery (ie. executing an HTTP call to the target endpoint).

The store is based on BoltDB, although we can certainly revisit this later.

* Some changes after Rafa's comment.

* more fixes